% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.to.graphs.R
\name{add.to.graphs}
\alias{add.to.graphs}
\title{-----------------------------------------------------------------------------
Augment grahs}
\usage{
add.to.graphs(graphs, sentinel, trans.genes, trans.cpgs, tfbs.ann)
}
\arguments{
\item{graphs}{list of graph objects to add nodes and edges to}

\item{sentinel}{character id of the sentinel SNP}

\item{trans.genes}{character vector of genes in the trans locus}

\item{trans.cpgs}{character vector of CpGs with trans meQTLs}

\item{tfbs.ann}{logical indicator matrix with CpGs in the rows and
transcription factors as columns}
}
\value{
list of graph objects with the nodes and edges added
}
\description{
Add TF to CpG edges and trans gene to SNP edges to graphs. This function
assumes that you have an object tfbs.ann in the environment. It is not
passed as argument (bad style) becuase it is too big (passing by value).
}
\author{
Matthias Heinig <matthias.heinig@helmholtz-muenchen.de>

-----------------------------------------------------------------------------
}
