% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eig.decomp.R
\name{eig.decomp}
\alias{eig.decomp}
\title{-----------------------------------------------------------------------------
Get eigenvalue decomposition}
\usage{
eig.decomp(M, n.eigs, sym)
}
\arguments{
\item{sym}{}
}
\value{

}
\description{
-----------------------------------------------------------------------------
Get eigenvalue decomposition
}
\author{
Matthias Heinig <matthias.heinig@helmholtz-muenchen.de>

-----------------------------------------------------------------------------
}
