% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.nodes.by.type.R
\name{get.nodes.by.type}
\alias{get.nodes.by.type}
\title{-----------------------------------------------------------------------------
Get nodes by type}
\usage{
get.nodes.by.type(graph, type)
}
\arguments{
\item{graph}{graphNEL (locus graph) object}

\item{type}{character indicating the node type}
}
\value{
Character vector containing nodes corresponding to the given type
}
\description{
Convenience function to retrieve nodes of a certain type from a graph
object using nodeData attributes
}
\author{
Matthias Heinig <matthias.heinig@helmholtz-muenchen.de>

-----------------------------------------------------------------------------
}
