% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_candidate_gene.R
\name{get_candidate_gene}
\alias{get_candidate_gene}
\title{-----------------------------------------------------------------------------
Get the candidate gene of a locus using the random walk approach.}
\usage{
get_candidate_gene(snp_id, snp_genes, trans_cpgs, ppi_graph,
  tfbs_annotation, gene_only = FALSE)
}
\arguments{
\item{snp_id}{Character id of the SNP in the locus}

\item{snp_genes}{Character vector of the genes encoded near the SNP}

\item{trans_cpgs}{Character vector of the trans CpGs associated
to the SNP}

\item{ppi_graph}{graphNEL object containing the PPIs (e.g. from STRING)}

\item{tfbs_annotation}{Logical indicator matrix with trans entities 
in the rows and transcription factors as columns (bincinding site
information of transcription factors at trans entities)}

\item{gene_only}{Logical whether to return only the identified candidate
gene or also the rest of the random walk results (default: FALSE)}
}
\value{
The candidate gene as determined by the random walk analysis
}
\description{
Applys the random walk approach to get the most likely SNP gene
from the list of provided SNP genes
}
\author{
Johann Hawe <johann.hawe@helmholtz-muenchen.de>

#'@examples
\dontrun{
data(ppi_string)
data(test_data)
data(tfbs_annotation)

candidate_gene <- get_candidate_gene(snp_id, 
                                     snp_genes, 
                                     trans_cpgs, 
                                     ppi_string,
                                     tfbs_annotation,
                                     gene_only = TRUE)
print(candidate_gene)
}
-----------------------------------------------------------------------------
}
