% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph2sparseMatrix.R
\name{graph2sparseMatrix}
\alias{graph2sparseMatrix}
\title{-----------------------------------------------------------------------------
Convert graph to a sparse matrix}
\usage{
graph2sparseMatrix(graph)
}
\arguments{
\item{graph}{The graph to be converted to sparse matrix representation}
}
\value{
The graph represented as a sparse matrix
}
\description{
-----------------------------------------------------------------------------
Convert graph to a sparse matrix
}
\author{
Matthias Heinig <matthias.heinig@helmholtz-muenchen.de>

-----------------------------------------------------------------------------
}
