% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min.node.weight.path.R
\name{min.node.weight.path}
\alias{min.node.weight.path}
\title{-----------------------------------------------------------------------------
Find the shortest path with minimal node weight.}
\usage{
\method{min}{node.weight.path}(g, weights, from, to)
}
\arguments{
\item{g}{The graphNEL object}

\item{weights}{The weights calculated for the nodes}

\item{from}{The start nodes from which to start the min-node-weight path}

\item{to}{The target nodes to which to calcaulate the min-node-weight path}
}
\description{
-----------------------------------------------------------------------------
Find the shortest path with minimal node weight.
}
\author{
Matthias Heinig <matthias.heinig@helmholtz-muenchen.de>

-----------------------------------------------------------------------------
}
