% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.propagation.network.R
\name{plot.propagation.network}
\alias{plot.propagation.network}
\title{-----------------------------------------------------------------------------
Convenience function to plot networks}
\usage{
plot.propagation.network(g, prop, cpgs, best.trans, trans.genes, sentinel,
  prefix, additional.genes = NULL, additional.bordercol = NULL)
}
\description{
-----------------------------------------------------------------------------
Convenience function to plot networks
}
\author{
Matthias Heinig <matthias.heinig@helmholtz-muenchen.de>

-----------------------------------------------------------------------------
}
