% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritization.table.R
\name{prioritization.table}
\alias{prioritization.table}
\title{-----------------------------------------------------------------------------
Prioritization summary table}
\usage{
prioritization.table(prioritize, full = FALSE, locus.graphs = NULL,
  subset = NULL)
}
\arguments{
\item{prioritize}{list of random walk analysis results generated with the
network_batch.R script}

\item{full}{boolean indicating whether all genes in the trans locus should
be reported (TRUE) or only the best (FALSE)}

\item{locus.graphs}{a list of graphNEL objects with more details on the
graphs used for analysis (for full summary)}

\item{subset}{list of subsets of trans genes for each locus this is useful
to recycle permutation runs on maximal intervals}
}
\value{
data.frame with the best candidate gene per locus
}
\description{
Extract summary information from the random walk analysis
}
\author{
Matthias Heinig <matthias.heinig@helmholtz-muenchen.de>

-----------------------------------------------------------------------------
}
