% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritize.by.randomwalk.R
\name{prioritze.by.randomwalk}
\alias{prioritze.by.randomwalk}
\title{-----------------------------------------------------------------------------
Candidate prioritization by random walk analysis}
\usage{
prioritze.by.randomwalk(locus.graph, largest.cc, string.nodes, tf.nodes,
  n.eigs = 500, mode = "propagation", return.g = FALSE)
}
\arguments{
\item{locus.graph}{graphNEL object as generated in network_preprocess.R}

\item{largest.cc}{graphNEL object with the largest connected component
of the protein - protein interaction network}

\item{string.nodes}{character vector with the names of the nodes in the PPI
network}

\item{tf.nodes}{character vector with the names of the transcription factors}

\item{n.eigs}{number of eigen vectors to use for the approximation}

\item{mode}{character "propagation" or "hitting.time"}

\item{return.g}{logical should the graph used for analysis be returned?}
}
\value{
list with sentinel, cpgs, tfs, trans.genes, prop,
  best.trans.gene, best.tf, g
}
\description{
-----------------------------------------------------------------------------
Candidate prioritization by random walk analysis
}
\author{
Matthias Heinig <matthias.heinig@helmholtz-muenchen.de>

-----------------------------------------------------------------------------
}
